requirements.txt
cachetools==4.2.4
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
google-api-core==1.31.6
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-translate==2.0.1
googleapis-common-protos==1.56.3
googletrans==4.0.0rc1
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==21.3
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
urllib3==1.26.9
